+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.1.13
+- Added custom method support for escaping string in processXmlMappings methods.
+- Added parameter in api_helper config file which defines custom escape string method of user.
+
 ## 0.1.12
 - Remove space from processXmlMappings methods.